c90 compliant

+++ 0x12-advanced_binary_search/0-advanced_binary.c
@@ -12,7 +12,8 @@ int advanced_binary(int *array, size_t size, int value)
 	if (array == NULL || size == 0 || value < array[0])
 		return (-1);
 
-	int high = (int)size - 1;
+
+	int high = size - 1;
 	int index;
 	int low = 0;
